Move wasm-related stuff to new coco-ui rust app
