+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["coco-core", "coco-vm"]
+members = ["coco-core", "coco-ui", "coco-vm"]
