+++ coco-ui/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "coco-ui"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+wasm-bindgen = "0.2.92"
+
+coco_core = { path = "../coco-core", package = "coco-core" }
+coco_vm = { path = "../coco-vm", package = "coco-vm" }
